+++ app/Http/Controllers/ExportCouponController.php
@@ -5,20 +5,29 @@ namespace App\Http\Controllers;
 use App\ExportCoupon;
 use Illuminate\Http\Request;
 use App\Repositories\ExportCouponRepository\IExportCouponRepository;
-
+use App\Helper\ICheckAction;
 class ExportCouponController extends Controller
 {
     private $exportcouponRepository;
+    private $checkAction;
 
-    public function __construct(IExportCouponRepository $exportcouponRepository)
+    public function __construct(ICheckAction $checkAction, IExportCouponRepository $exportcouponRepository)
     {
+        $this->checkAction = $checkAction;
         $this->exportcouponRepository = $exportcouponRepository;
     }
 
     public function index()
     {
-        return $this->exportcouponRepository->showIndex();
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->exportcouponRepository->checkRoleIndex($result);
+        if($check != 0){
+            return $this->exportcouponRepository->showIndex();
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
+
     public function viewExport(Request $request)
     {
         return $this->exportcouponRepository->showViewExport($request);
@@ -26,19 +35,36 @@ class ExportCouponController extends Controller
 
     public function export(Request $request)
     {
-        $this->exportcouponRepository->validateExport($request);
-        return $this->exportcouponRepository->exportMaterial($request);
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->exportcouponRepository->checkRoleStore($result);
+        if($check != 0){
+            $this->exportcouponRepository->validateExport($request);
+            return $this->exportcouponRepository->exportMaterial($request);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 
     public function exportSupplier(Request $request)
     {
-        return $this->exportcouponRepository->exportSupplier($request);
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->exportcouponRepository->checkRoleStore($result);
+        if($check != 0){
+            return $this->exportcouponRepository->exportSupplier($request);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
+
     public function getDetail($id)
     {
         return $this->exportcouponRepository->getDetailExport($id);
     }
 
+    public function getDetailExportSupplier($id)
+    {
+        return $this->exportcouponRepository->getDetailExportSupplier($id);
+    }
     public function printDetail($id)
     {
         return $this->exportcouponRepository->printDetailExport($id);
@@ -52,22 +78,29 @@ class ExportCouponController extends Controller
 
     public function destroyWarehouse(Request $request)
     {
-        return $this->exportcouponRepository->destroyWarehouse($request);
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->exportcouponRepository->checkRoleStore($result);
+        if($check != 0){
+            return $this->exportcouponRepository->destroyWarehouse($request);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 
     public function viewDestroyWarehouseCook($id)
     {
         return $this->exportcouponRepository->viewDestroyCook($id);
     }
+
     public function destroyWarehouseCook(Request $request)
     {
-        return $this->exportcouponRepository->destroyCook($request);
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->exportcouponRepository->checkRoleStore($result);
+        if($check != 0){
+            return $this->exportcouponRepository->destroyCook($request);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 
-    public function search(Request $request)
-    {
-        $count = ExportCoupon::selectRaw('count(code) as qty')->where('code','LIKE',"%{$request->searchExC}%")->value('qty');
-        $exportCoupons = ExportCoupon::where('code','LIKE',"%{$request->searchExC}%")->with('typeExport','detailExportCoupon')->get();
-        return view('exportcoupon.search',compact('count','exportCoupons'));
-    }
 }
